Fixed 413 issue with request body being too large
When uploading a photo on photo-tools.dynkarken.com, I'd get a 413 back from
the nginx server on the host. After a bit of investigation, I've learned that
the default limit to body size is 1mb[1]. To change this configuration, I'm
adding an nginx conf file as instructed by dokku[2], using the dokku standard
template from github master[3] and adding in my the `client_max_body_size`
directive under the `server` configuration.

[1]: https://nginx.org/en/docs/http/ngx_http_core_module.html#client_max_body_size
[2]: http://dokku.viewdocs.io/dokku/configuration/nginx/#customizing-the-nginx-configuration
[3]: https://github.com/dokku/dokku/blob/6b431378a15207e7a74579094496dc82990bb0bd/plugins/nginx-vhosts/templates/nginx.conf.sigil
Nginx