Adding ffmpeg buildpack
I've installed ffmpeg on my host to have it accessible by the python
application, but in hindsight this does nothing. Dokku runs everything
in containers and binaries installed on the host are not available to
applications. Instead I've added a buildpack adding ffmpeg to the
container using the following resources:

https://munich-made.com/2020/02/20200220073624-dokku-and-ffmpeg/
http://dokku.viewdocs.io/dokku/deployment/methods/buildpacks/