Upload video thumbnails
With this change for every video we upload, we also upload the first
frame of the video in all image sizes so that we can use it as a video
thumbnail.

I'm hopeful this might be useful on a phone which doesn't
automatically show an image for the video. It will certainly be useful
in the gallery view if a video is picked as cover image, in which case
the thumbnail can be used.