Serve images and videos from s3 using presigned urls
I'd serve files from ifany as a proxy for s3, but because of the way I
had coded up the proxy, it stalled when serving albums with large
videos. Instead of solving this issue, it felt easier to serve the files
from s3 directly and so here we are.

Documentation: https://boto3.amazonaws.com/v1/documentation/api/latest/guide/s3-presigned-urls.html