Adding password protection to photo-tools
I've naively assumed that I could leave the photo-tools page online
without password protection since nobody links to it. This turned out to
be a mistake and a lot of my album and gallery entries were corrupted by
what I think is several different attacks.

Some attackers tried SQL injection. Others polluted the data with
marketing materials. Either way I found it pretty naff and appalling,
but of course it's only a matter of someone stooping low enough for the
damage to be done, and there's plenty of people out there who are in
desperate situations.

As a solution I'm using flask-simplelogin to add password protection:
https://flask-simple-login.readthedocs.io/en/latest/index.html

It works just fine and I'm very pleased with the result.